Added a cache for geo data
